refactor: project structure refactor
This commit contains refactor of project structure by creating common
directory
